+++ src/main/java/com/api/sss/login/dto/request/BiometricLoginStartRequest.java
@@ -1,4 +1,4 @@
-package com.api.sss.login.dto;
+package com.api.sss.login.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
