+++ src/main/java/com/api/sss/login/dto/request/BiometricSignupRequest.java
@@ -1,4 +1,4 @@
-package com.api.sss.login.dto;
+package com.api.sss.login.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
