+++ src/main/java/com/api/sss/login/dto/request/RefreshTokenRequest.java
@@ -1,4 +1,4 @@
-package com.api.sss.login.dto;
+package com.api.sss.login.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
