+++ src/main/java/com/api/sss/login/dto/response/BiometricLoginStartResponse.java
@@ -1,4 +1,4 @@
-package com.api.sss.login.dto;
+package com.api.sss.login.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
@@ -6,7 +6,9 @@ import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class BiometricLoginStartResponse {
+public class
+
+BiometricLoginStartResponse {
 
 	@Schema(description = "서명 대상 Challenge 문자열", example = "f7a8f329-d2b4-4bd1-8c33-43008b87a212")
 	private String challenge;
