+++ src/main/java/com/api/sss/login/dto/response/RefreshTokenResponse.java
@@ -1,4 +1,4 @@
-package com.api.sss.login.dto;
+package com.api.sss.login.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
