+++ src/main/java/com/api/sss/login/service/LoginService.java
@@ -14,13 +14,13 @@ import org.springframework.stereotype.Service;
 import com.api.sss.config.exception.CustomException;
 import com.api.sss.config.exception.ErrorCode;
 import com.api.sss.jwt.JwtTokenProvider;
-import com.api.sss.login.dto.BiometricLoginStartRequest;
-import com.api.sss.login.dto.BiometricLoginStartResponse;
-import com.api.sss.login.dto.BiometricLoginVerifyRequest;
-import com.api.sss.login.dto.BiometricLoginVerifyResponse;
-import com.api.sss.login.dto.BiometricSignupRequest;
-import com.api.sss.login.dto.RefreshTokenRequest;
-import com.api.sss.login.dto.RefreshTokenResponse;
+import com.api.sss.login.dto.request.BiometricLoginStartRequest;
+import com.api.sss.login.dto.response.BiometricLoginStartResponse;
+import com.api.sss.login.dto.request.BiometricLoginVerifyRequest;
+import com.api.sss.login.dto.response.BiometricLoginVerifyResponse;
+import com.api.sss.login.dto.request.BiometricSignupRequest;
+import com.api.sss.login.dto.request.RefreshTokenRequest;
+import com.api.sss.login.dto.response.RefreshTokenResponse;
 import com.api.sss.member.entity.Member;
 import com.api.sss.member.repository.MemberRepository;
