+++ src/main/java/com/api/sss/model/controller/ModelController.java
@@ -1,30 +1,24 @@
 package com.api.sss.model.controller;
 
-import com.api.sss.model.dto.NewsRequest;
-import com.api.sss.model.dto.NewsResponse;
+import com.api.sss.model.dto.request.NewsRequest;
+import com.api.sss.model.dto.response.NewsResponse;
 import com.api.sss.model.service.ModelService;
-import org.springframework.http.HttpEntity;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
+
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.client.RestTemplate;
 
-import com.api.sss.config.exception.CustomException;
-import com.api.sss.config.exception.ErrorCode;
 import com.api.sss.config.response.dto.CustomResponse;
 import com.api.sss.config.response.dto.SuccessStatus;
-import com.api.sss.model.dto.ChatAskRequest;
-import com.api.sss.model.dto.ChatAskResponse;
+import com.api.sss.model.dto.request.ChatAskRequest;
+import com.api.sss.model.dto.response.ChatAskResponse;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.ExampleObject;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 
 import java.util.List;
